Create subdirectories in app/tests and move tests
