+++ src/web/server/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "ts-node-dev ./app/index.ts",
     "test": "jest",
-    "test-server": "npm test ./app/tests/*",
+    "test-server": "npm test ./app/tests/.*\\.ts",
     "dev": "concurrently \"npm run start\" \"cd ../client && npm start\"",
     "check": "tslint --config ./tslint.json --project ./tsconfig.json --force",
     "clean": "gts clean",
